[Fixes #9511] extrapolate FileUploadSizeLimit and UploadParallelismLimit limit from upload form
